Add word-transcription tests; change voicing model
To handle the phenomenon of forward devoicing of the [V] phoneme in
Polish, we change the model for changing voicing from one based on
individual letters to a new one which does two passes: one forward and
one backwards.

Lots of transcription unit tests are failing because the backward pass
is not implemented yet.